+++ examples/example.py
@@ -2,13 +2,17 @@
 # create a vector table, 
 # insert 10 vectors, 
 # and execute a vector similarity search.
+import sys
+import numpy as np
+import random
+sys.path.append(".")
 from milvus import Milvus, IndexType, MetricType
 import time
 
 # Milvus server IP address and port.
 # You may need to change _HOST and _PORT accordingly.
 _HOST = '127.0.0.1'
-_PORT = '19530'
+_PORT = '19530'  # default value
 
 
 def main():
@@ -18,16 +22,22 @@ def main():
     # You may need to change _HOST and _PORT accordingly
     param = {'host': _HOST, 'port': _PORT}
     status = milvus.connect(**param)
+    if status.OK():
+        print("Server connected.")
+    else:
+        print("Server connect fail.")
+        sys.exit(1)
 
     # Create table demo_table if it dosen't exist.
-    table_name = 'demo_table_01'
+    table_name = 'demo_table'
 
-    if not milvus.has_table(table_name):
+    status, ok = milvus.has_table(table_name)
+    if not ok:
         param = {
             'table_name': table_name,
             'dimension': 16,
-            'index_file_size': 1024,
-            'metric_type': MetricType.L2
+            'index_file_size': 1024,  # optional
+            'metric_type': MetricType.L2  # optional
         }
 
         milvus.create_table(param)
@@ -37,30 +47,38 @@ def main():
 
     # Describe demo_table
     _, table = milvus.describe_table(table_name)
+    print(table)
 
-    # 10 vectors with 16 dimension
-    vectors = [
-        [0.66, 0.01, 0.29, 0.64, 0.75, 0.94, 0.26, 0.79, 0.61, 0.11, 0.25, 0.50, 0.74, 0.37, 0.28, 0.63],
-        [0.77, 0.65, 0.57, 0.68, 0.29, 0.93, 0.17, 0.15, 0.95, 0.09, 0.78, 0.37, 0.76, 0.21, 0.42, 0.15],
-        [0.61, 0.38, 0.32, 0.39, 0.54, 0.93, 0.09, 0.81, 0.52, 0.30, 0.20, 0.59, 0.15, 0.27, 0.04, 0.37],
-        [0.33, 0.03, 0.87, 0.47, 0.79, 0.61, 0.46, 0.77, 0.62, 0.70, 0.85, 0.01, 0.30, 0.41, 0.74, 0.98],
-        [0.19, 0.80, 0.03, 0.75, 0.22, 0.49, 0.52, 0.91, 0.40, 0.91, 0.79, 0.08, 0.27, 0.16, 0.07, 0.24],
-        [0.44, 0.36, 0.16, 0.88, 0.30, 0.79, 0.45, 0.31, 0.45, 0.99, 0.15, 0.93, 0.37, 0.25, 0.78, 0.84],
-        [0.33, 0.37, 0.59, 0.66, 0.76, 0.11, 0.19, 0.38, 0.14, 0.37, 0.97, 0.50, 0.08, 0.69, 0.16, 0.67],
-        [0.68, 0.97, 0.20, 0.13, 0.30, 0.16, 0.85, 0.21, 0.26, 0.17, 0.81, 0.96, 0.18, 0.40, 0.13, 0.74],
-        [0.11, 0.26, 0.44, 0.91, 0.89, 0.79, 0.98, 0.91, 0.09, 0.45, 0.07, 0.88, 0.71, 0.35, 0.97, 0.41],
-        [0.17, 0.54, 0.61, 0.58, 0.25, 0.63, 0.65, 0.71, 0.26, 0.80, 0.28, 0.77, 0.69, 0.02, 0.63, 0.60],
-    ]
+    # 10000 vectors with 16 dimension
+    # element per dimension is float32 type
+    # vectors should be a 2-D array
+    vectors = [[random.random() for _ in range(16)] for _ in range(10000)]
+    # You can also use numpy to generate random vectors:
+    #     `vectors = np.random.rand(10000, 16).astype(np.float32).tolist()`
 
-    # Insert vectors into demo_table
-    status, ids = milvus.add_vectors(table_name=table_name, records=vectors)
+    # Insert vectors into demo_table, return status and vectors id list
+    status, ids = milvus.insert(table_name=table_name, records=vectors)
 
-    # Wait for 6 seconds, since Milvus server persist vector data every 5 seconds by default.
-    # You can set data persist interval in Milvus config file.
+    # Wait for 6 seconds, until Milvus server persist vector data.
     time.sleep(6)
 
     # Get demo_table row count
-    status, result = milvus.get_table_row_count(table_name)
+    status, result = milvus.count_table(table_name)
+
+    # create index of vectors, search more rapidly
+    index_param = {
+        'index_type': IndexType.IVFLAT,  # choice ivflat index
+        'nlist': 2048
+    }
+
+    # Create ivflat index in demo_table
+    # You can search vectors without creating index. however, Creating index help to
+    # search faster
+    status = milvus.create_index(table_name, index_param)
+
+    # describe index, get information of index
+    status, index = milvus.describe_index(table_name)
+    print(index)
 
     # Use the 3rd vector for similarity search
     query_vectors = [
@@ -77,13 +95,14 @@ def main():
     }
     status, results = milvus.search_vectors(**param)
 
-    if results[0][0].distance == 0.0 or results[0][0].id == ids[3]:
-        print('Query result is correct')
-    else:
-        print('Query result isn\'t correct')
+    if status.OK():
+        if results[0][0].distance == 0.0 or results[0][0].id == ids[3]:
+            print('Query result is correct')
+        else:
+            print('Query result isn\'t correct')
 
     # Delete demo_table
-    status = milvus.delete_table(table_name)
+    status = milvus.drop_table(table_name)
 
     # Disconnect from Milvus
     status = milvus.disconnect()