+++ milvus/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from .client.grpc_client import GrpcMilvus as Milvus
-from .client.grpc_client import Prepare
+from .client.prepare import Prepare
 from .client.types import IndexType, MetricType, Status
 from .client.exceptions import (
     ParamError,